refactor: add index.ts for apis, dtos
