+++ src/apis/dtos/index.ts
@@ -1,0 +1,2 @@
+export * as ScheduleBlockDto from './ScheduleBlockDto';
+export * as TimetableDto from './TimetableDto';
