+++ src/apis/index.ts
@@ -1,0 +1,2 @@
+export * as timetable from './timetable';
+export * as scheduleBlock from './scheduleBlock';
