+++ src/apis/scheduleBlock.ts
@@ -1,5 +1,5 @@
 import axios from 'utils/customAxios';
-import * as ScheduleBlockDto from './dtos/ScheduleBlock';
+import { ScheduleBlockDto } from './dtos';
 
 export const getScheduleBlocksByTimetableId = async (timetableId: string) => {
   const res = await axios.get<ScheduleBlockDto.Response>(`/timetables/${timetableId}/scheduleblocks`);
